refactor python scripts
