+++ scripts/check-lists.py
@@ -1,9 +1,10 @@
-import glob
 import itertools
-import json
 import operator
 from dataclasses import dataclass
 from functools import reduce
+
+from common_classes import Coin, ERC20Token
+from utils import read_json
 
 
 class CheckResult:
@@ -79,7 +80,7 @@ class Currency:
             return
 
         try:
-            price = ref.get_price(prices)
+            price = prices[ref.symbol]
         except Exception as e:
             yield Warning(self, f"No price: {e}")
             return
@@ -117,62 +118,10 @@ class Currency:
 
 
 @dataclass
-class Coin:
-    symbol: str
-    name: str
-    key: str
-    decimals: int
-    logo: str
-    website: str
-
-    def __str__(self):
-        return f"[{self.symbol}, COIN]"
-
-    def get_price(self, prices):
-        return prices[self.symbol]
-
-    def check(self):
-        if self.logo is None:
-            yield Warning(self, f"No logo")
-
-
-@dataclass
-class ERC20Token:
-    address: str
-    decimals: int
-    displaySymbol: str
-    logo: str
-    name: str
-    symbol: str
-    website: str
-
-    def __str__(self):
-        return f"[{self.symbol}, ERC20]"
-
-    def get_price(self, prices):
-        return prices[self.symbol]
-
-    def check(self):
-        if self.logo is None:
-            yield Warning(self, f"No logo")
-
-
-@dataclass
 class Chain:
     chain: str
     native: str
     tokens: str
-
-
-def read_json(path):
-    with open(path) as json_file:
-        return json.load(json_file)
-
-
-def multiread_json(base_dir, pattern):
-    for target in sorted(glob.glob(base_dir + pattern)):
-        key = target.replace(base_dir, '').partition("/")[0]
-        yield (key, read_json(target))
 
 
 def find_duplicates(items, key):
@@ -183,6 +132,11 @@ def find_duplicates(items, key):
 
 def compress_duplicates(duplicates):
     return [(symbol, [x.name for x in group]) for symbol, group in duplicates]
+
+
+def check_logo(coin):
+    if coin.logo is None:
+        yield Warning(coin, "No logo")
 
 
 def check_currencies(currencies, coins, erc20_tokens, chains, prices):
@@ -214,15 +168,15 @@ def check_currencies(currencies, coins, erc20_tokens, chains, prices):
             yield Error(currency, "Reference not found")
             continue
 
-        yield from itertools.chain(ref.check(), currency.check(ref, prices))
+        yield from itertools.chain(check_logo(ref), currency.check(ref, prices))
 
 
 def main():
-    coins = list(map(lambda x: Coin(**x), read_json("coins.json")))
-    erc20_tokens = list(map(lambda x: ERC20Token(**x), read_json("erc20-tokens.json")))
+    coins = list(map(lambda x: Coin.from_dict(x), read_json("coins.json")))
+    erc20_tokens = list(map(lambda x: ERC20Token.from_dict(x), read_json("erc20-tokens.json")))
     chains = list(map(lambda x: Chain(**x), read_json("chain/list.json")))
     chains = dict((c.native, read_json(c.tokens)) for c in chains)
-    chains = {k: list(map(lambda x: ERC20Token(**x), v)) for k, v in chains.items()}
+    chains = {k: list(map(lambda x: ERC20Token.from_dict(x), v)) for k, v in chains.items()}
 
     currencies = map(lambda x: Currency(**x), read_json("custody.json"))
