+++ scripts/grab-details-cryptocompare.py
@@ -1,80 +1,30 @@
 import datetime
-import glob
 import itertools
 import json
 import os
 import re
-from dataclasses import dataclass
 from typing import Union, Optional, Dict
 
 import requests
+
+from common_classes import Coin, ERC20Token
+from utils import read_json, multiread_json
 
 Timestamp = Union[datetime.datetime, datetime.date, int, float]
 
 DATA_URL = 'https://min-api.cryptocompare.com/data/all/coinlist'
 
 
-@dataclass
-class Coin:
-    symbol: str
-    name: str
-    key: str
-    decimals: int
-    logo: str
-
-    def __str__(self):
-        return f"[{self.symbol}, COIN]"
-
-    def get_price(self, prices):
-        return prices[self.symbol]
-
-    def check(self):
-        if self.logo is None:
-            yield Warning(self, f"No logo")
-
-
-@dataclass
-class ERC20Token:
-    address: str
-    decimals: int
-    displaySymbol: str
-    logo: str
-    name: str
-    symbol: str
-    website: str
-
-    def __str__(self):
-        return f"[{self.symbol}, ERC20]"
-
-    def get_price(self, prices):
-        return prices[self.symbol]
-
-    def check(self):
-        if self.logo is None:
-            yield Warning(self, f"No logo")
-
-
-def read_json(path):
-    with open(path) as json_file:
-        return json.load(json_file)
-
-
-def multiread_json(base_dir, pattern):
-    for target in sorted(glob.glob(base_dir + pattern)):
-        key = target.replace(base_dir, '').partition("/")[0]
-        yield key, read_json(target)
-
-
 def compress_duplicates(duplicates):
     return [(symbol, [x.name for x in group]) for symbol, group in duplicates]
 
 
-def _query_cryptocompare(url: str, errorCheck: bool = True, api_key: str = None) -> Optional[Dict]:
+def _query_cryptocompare(url: str, error_check: bool = True, api_key: str = None) -> Optional[Dict]:
     """
     Query the url and return the result or None on failure.
     :param url: the url
-    :param errorCheck: run extra error checks (default: True)
-    :returns: respones, or nothing if errorCheck=True
+    :param error_check: run extra error checks (default: True)
+    :returns: response, or nothing if errorCheck=True
     :api_key: optional, if you want to add an API Key
     """
 
@@ -84,10 +34,10 @@ def _query_cryptocompare(url: str, errorCheck: bool = True, api_key: str = None)
     try:
         response = requests.get(url + api_key_parameter).json()
     except Exception as e:
-        print('Error getting coin information. %s' % str(e))
+        print(f'Error getting coin information. {str(e)}')
         return None
-    if errorCheck and (response.get('Response') == 'Error'):
-        print('[ERROR] %s' % response.get('Message'))
+    if error_check and (response.get('Response') == 'Error'):
+        print(f'[ERROR] {response.get("Message")}')
         return None
     return response
 
@@ -102,12 +52,12 @@ def filter_desc(line):
 
 
 def main():
-    coins = list(map(lambda x: Coin(**x), read_json("coins.json")))
-    erc20_tokens = list(map(lambda x: ERC20Token(**x), read_json("erc20-tokens.json")))
+    coins = list(map(lambda x: Coin.from_dict(x), read_json("coins.json")))
+    erc20_tokens = list(map(lambda x: ERC20Token.from_dict(x), read_json("erc20-tokens.json")))
     crypto_compare_data = _query_cryptocompare(DATA_URL)
     crypto_compare_data = crypto_compare_data['Data']
     chains = dict(multiread_json("chain/", "*/tokens.json"))
-    chains = {k: list(map(lambda x: ERC20Token(**x), v)) for k, v in chains.items()}
+    chains = {k: list(map(lambda x: ERC20Token.from_dict(x), v)) for k, v in chains.items()}
 
     map_symbols = set(itertools.chain.from_iterable(
         [map(lambda x: x.symbol, itertools.chain(*chains.values())),
@@ -115,7 +65,7 @@ def main():
          map(lambda x: x.symbol, erc20_tokens)]))
     dic_list = dict()
     no_symbol = list()
-    objectArray = []
+    object_array = []
     for sym in map_symbols:
         if sym in crypto_compare_data:
             crypto_details = crypto_compare_data[sym]
@@ -127,12 +77,12 @@ def main():
             desc_text = filter_desc(crypto_details['Description'])
             dic_list[sym] = desc_text
             desc['description'] = desc_text
-            objectArray.append(desc)
+            object_array.append(desc)
         else:
             no_symbol.append(sym)
 
     write_json(dic_list, './description/en.json')
-    write_json(objectArray, './description/info.json')
+    write_json(object_array, './description/info.json')
     print('Warning No Description Found For These:', no_symbol)
